+++ packages/frontend/src/pages/scaling/activity/view/ScalingActivityView.tsx
@@ -2,9 +2,9 @@ import React from 'react'
 
 import { ScalingFilters } from '../../../../components/table/filters/ScalingFilters'
 import { getScalingRowProps } from '../../../../components/table/props/getScalingRowProps'
-import { getScalingActivityColumnsConfig } from '../../../../components/table/props/getScalingTableColumnsConfig'
 import { TableView } from '../../../../components/table/TableView'
 import { RowConfig } from '../../../../components/table/types'
+import { getScalingActivityColumnsConfig } from '../props/getScalingActivityColumnsConfig'
 import { ActivityViewEntry } from '../types'
 export interface ScalingActivityViewProps {
   items: ActivityViewEntry[]
