+++ packages/frontend/src/pages/scaling/liveness/view/ScalingLivenessView.tsx
@@ -3,9 +3,9 @@ import React from 'react'
 import { ScalingLegend } from '../../../../components/ScalingLegend'
 import { ScalingFilters } from '../../../../components/table/filters/ScalingFilters'
 import { getScalingRowProps } from '../../../../components/table/props/getScalingRowProps'
-import { getScalingLivenessColumnsConfig } from '../../../../components/table/props/getScalingTableColumnsConfig'
 import { TableView } from '../../../../components/table/TableView'
 import { RowConfig } from '../../../../components/table/types'
+import { getScalingLivenessColumnsConfig } from '../props/getScalingLivenessColumnsConfig'
 import { ScalingLivenessViewEntry } from '../types'
 import { LivenessTimeRangeControls } from './LivenessTimeRangeControls'
