+++ packages/frontend/src/pages/scaling/risk/view/ScalingRiskView.tsx
@@ -4,10 +4,10 @@ import { ActiveIcon } from '../../../../components/icons/symbols/ActiveIcon'
 import { ArchivedIcon } from '../../../../components/icons/symbols/ArchivedIcon'
 import { ScalingFilters } from '../../../../components/table/filters/ScalingFilters'
 import { getScalingRowProps } from '../../../../components/table/props/getScalingRowProps'
-import { getScalingRiskColumnsConfig } from '../../../../components/table/props/getScalingTableColumnsConfig'
 import { TableView } from '../../../../components/table/TableView'
 import { RowConfig } from '../../../../components/table/types'
 import { Tabs } from '../../../../components/Tabs'
+import { getScalingRiskColumnsConfig } from '../props/getScalingRiskColumnsConfig'
 import { ScalingRiskViewEntry } from '../types'
 
 export interface ScalingRiskViewProps {
